pyproject.toml
[project]
name = "ir-codesearchnet-bm25-codebert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.8.4",
    "faiss-cpu>=1.8.0",
    "numpy>=2.0.0",
    "torch>=2.4.0",
    "tqdm>=4.67.3",
    "transformers>=4.45.0",
]
